sessions: show New Session button in titlebar when sidebar is hidden
Surface the compact "New" session affordance in the titlebar's left
section when the sessions list (sidebar) is hidden, mirroring the button
in the sessions sidebar header.

This is done with a single action (NEW_SESSION_ACTION_ID) contributed to
both Menus.SidebarSessionsHeader and Menus.TitleBarLeftLayout, rendered
by one shared NewSessionActionViewItem registered for both menus via
IActionViewItemService. The view item inlines the compact "New" pill
(label + live keybinding hint + hover) so both surfaces render the exact
same widget. The titlebar entry is gated to only show when the sidebar is
hidden, since the list has its own button.

The sidebar header's hand-placed button is removed in favor of the
menu-driven action view item, and the button CSS now lives in an
unscoped media/newSessionActionViewItem.css so it renders identically
wherever it is mounted.
